Add in picard metrics
